Working for unit testing, completed Repository unit testing

+++ app/src/main/java/com/mpg/nagarro/deviceinventorymgmt/ui/worker/CleanupWorker.kt
@@ -5,9 +5,7 @@ import android.util.Log
 import androidx.work.Worker
 import androidx.work.WorkerParameters
 import com.mpg.nagarro.deviceinventorymgmt.util.OUTPUT_PATH
-import com.mpg.nagarro.deviceinventorymgmt.util.makeStatusNotification
-import com.mpg.nagarro.deviceinventorymgmt.util.sleep
-import timber.log.Timber
+import com.mpg.nagarro.deviceinventorymgmt.util.Utils
 import java.io.File
 
 /**
@@ -17,10 +15,6 @@ import java.io.File
 class CleanupWorker(ctx: Context, params: WorkerParameters) : Worker(ctx, params) {
     private val TAG = "CleanupWorker"
     override fun doWork(): Result {
-        // Makes a notification when the work starts and slows down the work so that
-        // it's easier to see each WorkRequest start, even on emulated devices
-        makeStatusNotification("Cleaning up 30days older Backup files", applicationContext)
-        sleep()
         Log.i(TAG, "onCreate= : doWork====>")
         return try {
             val outputDirectory = File(applicationContext.filesDir, OUTPUT_PATH)
@@ -30,15 +24,23 @@ class CleanupWorker(ctx: Context, params: WorkerParameters) : Worker(ctx, params
                     for (entry in entries) {
                         val name = entry.name
                         if (name.isNotEmpty() && name.endsWith(".csv")) {
-                            val deleted = entry.delete()
-                            Log.i(TAG,"Deleted $name - $deleted")
+                            val startIn = name.indexOf("_")
+                            val timeInSec = name.substring(startIn, (name.length - 4)).toLong()
+
+                            val numDays = Utils.getDifferenceInDays(timeInSec)
+
+                            if (numDays >= 30) {
+                                val deleted = entry.delete()
+                                Log.i(TAG, "Deleted $name - $deleted")
+                            }
+                            Log.i(TAG, "name: Cleanup worker = $name")
                         }
                     }
                 }
             }
             Result.success()
         } catch (exception: Exception) {
-            Timber.e(exception)
+            //Timber.e(exception)
             Result.failure()
         }
     }
